Initial run of immaculate
- Hasn't fully been run, certain files cause exceptions

+++ src/main/java/com/simibubi/create/content/equipment/tool/CardboardSwordItem.java
@@ -1,7 +1,5 @@
 package com.simibubi.create.content.equipment.tool;
 
-import java.util.function.Consumer;
-
 import com.simibubi.create.AllItems;
 import com.simibubi.create.AllSoundEvents;
 import com.simibubi.create.foundation.item.render.SimpleCustomRenderer;
@@ -26,7 +24,6 @@ import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.item.enchantment.ItemEnchantments;
-
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
 import net.neoforged.bus.api.EventPriority;
@@ -37,92 +34,102 @@ import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
 import net.neoforged.neoforge.event.entity.player.AttackEntityEvent;
 import net.neoforged.neoforge.event.entity.player.PlayerInteractEvent;
 
+import java.util.function.Consumer;
+
 @EventBusSubscriber
 public class CardboardSwordItem extends SwordItem {
 
-	public CardboardSwordItem(Properties pProperties) {
-		super(AllToolMaterials.CARDBOARD, pProperties);
-	}
+    public CardboardSwordItem(Properties pProperties) {
+        super(AllToolMaterials.CARDBOARD, pProperties);
+    }
 
-	@Override
-	public boolean supportsEnchantment(ItemStack stack, Holder<Enchantment> enchantment) {
-		return enchantment.getKey() == Enchantments.KNOCKBACK;
-	}
+    @Override
+    public boolean supportsEnchantment(ItemStack stack, Holder<Enchantment> enchantment) {
+        return enchantment.getKey() == Enchantments.KNOCKBACK;
+    }
 
-	@Override
-	public boolean isBookEnchantable(ItemStack stack, ItemStack book) {
-		ItemEnchantments enchants = book.getOrDefault(DataComponents.STORED_ENCHANTMENTS, ItemEnchantments.EMPTY);
-		for (Holder<Enchantment> enchantment : enchants.keySet()) {
-			if (enchantment.getKey() != Enchantments.KNOCKBACK)
-				return false;
-		}
-		return true;
-	}
+    @Override
+    public boolean isBookEnchantable(ItemStack stack, ItemStack book) {
+        ItemEnchantments enchants =
+                book.getOrDefault(DataComponents.STORED_ENCHANTMENTS, ItemEnchantments.EMPTY);
+        for (Holder<Enchantment> enchantment : enchants.keySet()) {
+            if (enchantment.getKey() != Enchantments.KNOCKBACK) return false;
+        }
+        return true;
+    }
 
-	@SubscribeEvent
-	public static void cardboardSwordsMakeNoiseOnClick(PlayerInteractEvent.LeftClickBlock event) {
-		ItemStack itemStack = event.getItemStack();
-		if (!AllItems.CARDBOARD_SWORD.isIn(itemStack))
-			return;
-		if (event.getAction() != PlayerInteractEvent.LeftClickBlock.Action.START)
-			return;
-		if (event.getSide() == LogicalSide.CLIENT)
-			AllSoundEvents.CARDBOARD_SWORD.playAt(event.getLevel(), event.getPos(), 0.5f, 1.85f, false);
-		else
-			AllSoundEvents.CARDBOARD_SWORD.play(event.getLevel(), event.getEntity(), event.getPos(), 0.5f, 1.85f);
-	}
+    @SubscribeEvent
+    public static void cardboardSwordsMakeNoiseOnClick(PlayerInteractEvent.LeftClickBlock event) {
+        ItemStack itemStack = event.getItemStack();
+        if (!AllItems.CARDBOARD_SWORD.isIn(itemStack)) return;
+        if (event.getAction() != PlayerInteractEvent.LeftClickBlock.Action.START) return;
+        if (event.getSide() == LogicalSide.CLIENT)
+            AllSoundEvents.CARDBOARD_SWORD.playAt(
+                    event.getLevel(), event.getPos(), 0.5f, 1.85f, false);
+        else
+            AllSoundEvents.CARDBOARD_SWORD.play(
+                    event.getLevel(), event.getEntity(), event.getPos(), 0.5f, 1.85f);
+    }
 
-	// We set priority to highest just so we catch this before anyone does anything else
-	@SubscribeEvent(priority = EventPriority.HIGHEST)
-	public static void cardboardSwordsCannotHurtYou(AttackEntityEvent event) {
-		Player attacker = event.getEntity();
-		if (!(event.getTarget() instanceof LivingEntity target) || target.getType().is(EntityTypeTags.ARTHROPOD))
-			return;
-		ItemStack stack = attacker.getItemInHand(InteractionHand.MAIN_HAND);
-		if (!(AllItems.CARDBOARD_SWORD.isIn(stack)))
-			return;
+    // We set priority to highest just so we catch this before anyone does anything else
+    @SubscribeEvent(priority = EventPriority.HIGHEST)
+    public static void cardboardSwordsCannotHurtYou(AttackEntityEvent event) {
+        Player attacker = event.getEntity();
+        if (!(event.getTarget() instanceof LivingEntity target)
+                || target.getType().is(EntityTypeTags.ARTHROPOD)) return;
+        ItemStack stack = attacker.getItemInHand(InteractionHand.MAIN_HAND);
+        if (!(AllItems.CARDBOARD_SWORD.isIn(stack))) return;
 
-		AllSoundEvents.CARDBOARD_SWORD.playFrom(attacker, 0.75f, 1.85f);
+        AllSoundEvents.CARDBOARD_SWORD.playFrom(attacker, 0.75f, 1.85f);
 
-		event.setCanceled(true);
+        event.setCanceled(true);
 
-		// Reference player.attack()
-		// This section replicates knockback behaviour without hurting the target
+        // Reference player.attack()
+        // This section replicates knockback behaviour without hurting the target
 
-		float knockbackStrength = (float) (attacker.getAttributeValue(Attributes.ATTACK_KNOCKBACK) + 2);
-		if (attacker.level() instanceof ServerLevel serverLevel)
-			knockbackStrength = EnchantmentHelper.modifyKnockback(serverLevel, stack, target, serverLevel.damageSources().playerAttack(attacker), knockbackStrength);
-		if (attacker.isSprinting() && attacker.getAttackStrengthScale(0.5f) > 0.9f)
-			++knockbackStrength;
+        float knockbackStrength =
+                (float) (attacker.getAttributeValue(Attributes.ATTACK_KNOCKBACK) + 2);
+        if (attacker.level() instanceof ServerLevel serverLevel)
+            knockbackStrength = EnchantmentHelper.modifyKnockback(
+                    serverLevel,
+                    stack,
+                    target,
+                    serverLevel.damageSources().playerAttack(attacker),
+                    knockbackStrength);
+        if (attacker.isSprinting() && attacker.getAttackStrengthScale(0.5f) > 0.9f)
+            ++knockbackStrength;
 
-		if (knockbackStrength <= 0)
-			return;
+        if (knockbackStrength <= 0) return;
 
-		float yRot = attacker.getYRot();
-		knockback(target, knockbackStrength, yRot);
+        float yRot = attacker.getYRot();
+        knockback(target, knockbackStrength, yRot);
 
-		boolean targetIsPlayer = target instanceof Player;
-		MobCategory targetType = target.getClassification(false);
+        boolean targetIsPlayer = target instanceof Player;
+        MobCategory targetType = target.getClassification(false);
 
-		if (target instanceof ServerPlayer sp)
-			CatnipServices.NETWORK.sendToClient(sp, new KnockbackPacket(yRot, (float) knockbackStrength));
+        if (target instanceof ServerPlayer sp)
+            CatnipServices.NETWORK.sendToClient(sp, new KnockbackPacket(yRot, knockbackStrength));
 
-		if ((targetType == MobCategory.MISC || targetType == MobCategory.CREATURE) && !targetIsPlayer)
-			target.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SLOWDOWN, 60, 9, true, false, false));
+        if ((targetType == MobCategory.MISC || targetType == MobCategory.CREATURE)
+                && !targetIsPlayer)
+            target.addEffect(
+                    new MobEffectInstance(MobEffects.MOVEMENT_SLOWDOWN, 60, 9, true, false, false));
 
-		attacker.setDeltaMovement(attacker.getDeltaMovement()
-			.multiply(0.6D, 1.0D, 0.6D));
-		attacker.setSprinting(false);
-	}
+        attacker.setDeltaMovement(attacker.getDeltaMovement().multiply(0.6D, 1.0D, 0.6D));
+        attacker.setSprinting(false);
+    }
 
-	public static void knockback(LivingEntity target, double knockbackStrength, float yRot) {
-		target.stopRiding();
-		target.knockback(knockbackStrength * 0.5F, Mth.sin(yRot * Mth.DEG_TO_RAD), -Mth.cos(yRot * Mth.DEG_TO_RAD));
-	}
+    public static void knockback(LivingEntity target, double knockbackStrength, float yRot) {
+        target.stopRiding();
+        target.knockback(
+                knockbackStrength * 0.5F,
+                Mth.sin(yRot * Mth.DEG_TO_RAD),
+                -Mth.cos(yRot * Mth.DEG_TO_RAD));
+    }
 
-	@Override
-	@OnlyIn(Dist.CLIENT)
-	public void initializeClient(Consumer<IClientItemExtensions> consumer) {
-		consumer.accept(SimpleCustomRenderer.create(this, new CardboardSwordItemRenderer()));
-	}
+    @Override
+    @OnlyIn(Dist.CLIENT)
+    public void initializeClient(Consumer<IClientItemExtensions> consumer) {
+        consumer.accept(SimpleCustomRenderer.create(this, new CardboardSwordItemRenderer()));
+    }
 }
